% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edan_content_getcontent.R
\name{edan_content_getcontent}
\alias{edan_content_getcontent}
\title{Get content}
\usage{
edan_content_getcontent(
  url = NA,
  id = NA,
  linkedContent = TRUE,
  AppID = NA,
  AppKey = NA,
  returnjson = FALSE
)
}
\arguments{
\item{url}{The url of the content}

\item{id}{The id of the content.}

\item{linkedContent}{A linked content option. Default value: TRUE}

\item{AppID}{AppID used for authentication}

\item{AppKey}{Key for the AppID used for authentication}

\item{returnjson}{If FALSE (default), converts the answer from EDAN to a list. If TRUE, returns the answer as JSON.}
}
\value{
List or JSON with the details of the item.
}
\description{
This endpoint retrieves unique atom level content. Either the id and url parameter is required but not both.
}
\details{
All queries require valid EDAN credentials (AppID and AppKey). Consult the EDAN Docs (https://edandoc.si.edu/) for instructions on how to obtain a set of credentials.
}
