% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edan_ead_search.R
\name{edan_ead_search}
\alias{edan_ead_search}
\title{EAD Search}
\usage{
edan_ead_search(
  query = NA,
  fqs = NA,
  AppID = NA,
  AppKey = NA,
  rows = 10,
  start = 0,
  returnjson = FALSE
)
}
\arguments{
\item{query}{The search term. Default value: *:*}

\item{fqs}{A JSON array of filter query params.}

\item{AppID}{AppID used for authentication}

\item{AppKey}{Key for the AppID used for authentication}

\item{rows}{Number of rows to return, max is 100.}

\item{start}{Start number, to use with rows}

\item{returnjson}{Boolean to return the answer as JSON, otherwise as a list.}
}
\value{
List or JSON with the results.
}
\description{
Returns structured data rows relating to ead type of content.
}
\details{
All queries require valid EDAN credentials (AppID and AppKey). Consult the EDAN Docs (https://edandoc.si.edu/) for instructions on how to obtain a set of credentials.
}
