% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edan_metadata_suggestions.R
\name{edan_metadata_suggestions}
\alias{edan_metadata_suggestions}
\title{Get Suggestions}
\usage{
edan_metadata_suggestions(
  suggest = "*:*",
  AppID = NA,
  AppKey = NA,
  returnjson = FALSE
)
}
\arguments{
\item{suggest}{The search term. Default value: *:*}

\item{AppID}{AppID used for authentication}

\item{AppKey}{Key for the AppID used for authentication}

\item{returnjson}{Boolean to return the answer as JSON, otherwise as a list.}
}
\value{
List or JSON with the results.
}
\description{
Returns suggested content for a given search term.
}
\details{
All queries require valid EDAN credentials (AppID and AppKey). Consult the EDAN Docs (https://edandoc.si.edu/) for instructions on how to obtain a set of credentials.
}
