% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContentEDAN.R
\name{getContentEDAN}
\alias{getContentEDAN}
\title{Get details of an item}
\usage{
getContentEDAN(itemID, AppID, AppKey, returnjson = FALSE)
}
\arguments{
\item{itemID}{ID of the item}

\item{AppID}{AppID used for authentication}

\item{AppKey}{Key for the AppID used for authentication}

\item{returnjson}{If FALSE (default), converts the answer from EDAN to a list. If TRUE, returns the answer as json.}
}
\value{
List or JSON with the details of the item.
}
\description{
Get the details of an item from the EDAN API.
}
