% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjectList.R
\name{getObjectList}
\alias{getObjectList}
\title{Retrieve object lists from EDAN using the EDAN API.}
\usage{
getObjectList(listID, AppID, AppKey, returnjson = FALSE)
}
\arguments{
\item{listID}{The string which most commonly can be found attached to object groups. IE https://americanhistory.si.edu/collections/object-groups/tokuno-gift has the object list with the id "p2a-1502402857224-1502899539095-0".}

\item{AppID}{AppID used for authentication}

\item{AppKey}{Key for the AppID used for authentication}

\item{returnjson}{If FALSE (default), converts the answer from EDAN to a list. If TRUE, returns the answer as JSON.}
}
\description{
All queries require valid EDAN credentials (AppID and AppKey). Consult the EDAN Docs (https://edandoc.si.edu/) for instructions on how to obtain a set of credentials.
Same library requirements as getContentEDAN function.
library("devtools", lib.loc="~/R/R-3.5.1/library")
}
