% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAPI.R
\name{queryAPI}
\alias{queryAPI}
\title{Prepares the query to the EDAN API}
\usage{
queryAPI(
  AppID = NA,
  AppKey = NA,
  QueryParameters = NA,
  api_url = NA,
  rows = 10,
  start = 0
)
}
\arguments{
\item{AppID}{AppID used for authentication}

\item{AppKey}{Key for the AppID used for authentication}

\item{QueryParameters}{query params}

\item{api_url}{URL of the route}

\item{rows}{Number of rows to return, max is 100.}

\item{start}{Start number, to use with rows}
}
\value{
Result of the GET request
}
\description{
Prepares the query to the EDAN API
}
