% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContentEDAN.R
\name{getContentEDAN}
\alias{getContentEDAN}
\title{Get details of an item}
\usage{
getContentEDAN(itemID, AppID, AppKey)
}
\arguments{
\item{itemID}{ID of the item}

\item{AppID}{AppID}

\item{AppKey}{Key for the App ID}
}
\value{
JSON with the details of the item.
}
\description{
Function to check the details of an item from the EDAN API.
}
