% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchEDAN.R
\name{searchEDAN}
\alias{searchEDAN}
\title{Searches EDAN}
\usage{
searchEDAN(query, AppID, AppKey, rows = 10, start = 0)
}
\arguments{
\item{query}{Query to run}

\item{AppID}{AppID}

\item{AppKey}{Key for the App ID}

\item{rows}{Number of rows to return, max is 100}

\item{start}{Start number, to use with rows}
}
\value{
JSON with the results.
}
\description{
Function to search the EDAN API.
}
