% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchEDAN.R
\name{searchEDAN}
\alias{searchEDAN}
\title{Searches EDAN}
\usage{
searchEDAN(query, AppID, AppKey, rows = 10, start = 0, returnjson = FALSE)
}
\arguments{
\item{query}{Query to run}

\item{AppID}{AppID used for authentication}

\item{AppKey}{Key for the AppID used for authentication}

\item{rows}{Number of rows to return, max is 100.}

\item{start}{Start number, to use with rows}

\item{returnjson}{If FALSE (default), converts the answer from EDAN to a list. If TRUE, returns the answer as json.}
}
\value{
List or JSON with the results.
}
\description{
Search metadata in the EDAN API.
}
